Fixed Mypy (#341)
* fixed lint

* make mypy happy

---------

+++ train.py
@@ -180,8 +180,8 @@ def main(args: TrainingArgs) -> None:  # noqa: C901
     # Initialize trainer
     supported_quantizations = ["nf4", "nf4-dq", "fp4", "fp4-dq", "int8", "int8-training"]
     if args.precision in supported_quantizations:
-        args.plugins = BitsandbytesPrecision(mode=args.precision)
-        args.precision = "bf16-true"
+        args.plugins = BitsandbytesPrecision(mode=args.precision)  # type: ignore
+        args.precision = "16-true"
 
     trainer = Trainer(
         num_sanity_val_steps=0,
@@ -194,14 +194,14 @@ def main(args: TrainingArgs) -> None:  # noqa: C901
         logger=wandb_logger,
         deterministic=args.force_deterministic,
         callbacks=callbacks,
-        precision=args.precision,
+        precision=args.precision,  # type: ignore
         gradient_clip_val=args.grad_clip,
         log_every_n_steps=24,
         # accumulate_grad_batches=args.gradient_accumulation_steps,
         fast_dev_run=args.fast_dev_run,
         profiler=args.profiler,
         inference_mode=not args.compile,  # inference_mode for val/test and PyTorch 2.0 compiler don't like each other
-        plugins=args.plugins,
+        plugins=args.plugins,  # type: ignore
         # reload_dataloaders_every_n_epochs=1,
     )
